Remove core and util warnings (#8239)
* Fix some signed/unsigned mismatches

* Skip unused result warning

* Explict fallthrough for murmur hash

* Enable aligned new support to eliminate warning

* Switch to int instead of unsigned in some cases

+++ caffe2/core/plan_executor.cc
@@ -437,7 +437,7 @@ bool ExecuteStepRecursive(ExecutionStepWrapper& stepWrapper) {
         if (step.has_num_concurrent_instances()) {
           numThreads *= step.num_concurrent_instances();
         }
-        for (int64_t i = 0; i < numThreads; ++i) {
+        for (size_t i = 0; i < numThreads; ++i) {
           threads.emplace_back(worker);
         }
         for (auto& thread : threads) {
